Changed several components to refresh account details autonomously
