+++ Nostrid.Core/Data/FeedService.cs
@@ -132,7 +132,6 @@ public class FeedService
             if (rootTrees.Exists(tw.Id)) continue;
             var root = rootTrees.Find(tw.ReplyToId);
             var newTree = new NoteTree(tw);
-            newTree.Details = accountService.GetAccountDetails(tw.PublicKey);
             if (root != null)
             {
                 root.Children.Add(newTree);
