+++ Nostrid.Core/Shared/FollowAccount.razor
@@ -7,26 +7,26 @@
 
 @code {
     [Parameter, EditorRequired]
-    public Account Account { get; set; } = null!;
+    public string AccountId { get; set; }
 }
 
-<div role="button" @onclick="@(() => navigationManager.NavigateTo($"account/{Account.Id}"))" class="d-flex flex-row align-items-start p-3 bg-tertiary-hover text-decoration-none">
+<div role="button" @onclick="@(() => navigationManager.NavigateTo($"account/{AccountId}"))" class="d-flex flex-row align-items-start p-3 bg-tertiary-hover text-decoration-none">
     <div class="d-flex flex-shrink-0">
-        <AccountPicture Details="@Account.Details" Id="@Account.Id" Size="48" />
+        <AccountPicture Id="@AccountId" Size="48" />
     </div>
     <div class="d-flex flex-column flex-grow-1 ms-3">
         <div class="d-flex flex-row align-items-center">
             <div class="fs-5 fw-bolder lh-1 flex-grow-1">
-                <AccountName Details="@Account.Details" Id="@Account.Id" />
+                <AccountName Id="@AccountId" />
             </div>
-            @if (IsMainAccountLoaded && accountService.MainAccount?.Id != Account.Id)
+            @if (IsMainAccountLoaded && accountService.MainAccount?.Id != AccountId)
             {
-                <button class="btn btn-sm @(!accountService.IsFollowing(Account.Id) ? "btn-primary" : "btn-danger")" @onclick="@(async () => await FollowUnfollow(Account.Id))" @onclick:stopPropagation="true">
-                    @(!accountService.IsFollowing(Account.Id) ? "Follow" : "Unfollow")
+                <button class="btn btn-sm @(!accountService.IsFollowing(AccountId) ? "btn-primary" : "btn-danger")" @onclick="@(async () => await FollowUnfollow(AccountId))" @onclick:stopPropagation="true">
+                    @(!accountService.IsFollowing(AccountId) ? "Follow" : "Unfollow")
                 </button>
             }
         </div>
-        <div class="fs-7 mt-1 lh-1 text-body-tertiary text-break">@ByteTools.PubkeyToNpub(Account.Id)</div>
+        <div class="fs-7 mt-1 lh-1 text-body-tertiary text-break">@ByteTools.PubkeyToNpub(AccountId)</div>
     </div>
 </div>
