+++ Nostrid.Core/Shared/NavMenu.razor
@@ -13,7 +13,7 @@
 <nav class="navbar navbar-light bg-body-tertiary fixed-top d-md-none m-0 px-2 py-1 border-bottom bg-opacity-95">
 	<div class="flex-grow-1 d-flex flex-row justify-content-between align-items-center">
 		<NavLink class="nav-link" href="#">
-			<AccountPicture Details="@accountService.MainAccount?.Details" Id="@accountService.MainAccount?.Id" Size="35"></AccountPicture>
+			<AccountPicture Id="@accountService.MainAccount?.Id" Size="35"></AccountPicture>
 		</NavLink>
 		<div class="flex-grow-1 text-center">
 			<a class="navbar-brand m-0" href="/">Nostrid</a>
