+++ Nostrid.Core/Shared/NoteViewer.razor
@@ -43,7 +43,7 @@
     public Action<Event>? OnRepost { get; set; }
 
     [Parameter]
-    public Action<AccountDetails>? OnPay { get; set; }
+    public Action<string>? OnPay { get; set; }
 
     [Parameter]
     public Action<Event>? OnDelete { get; set; }
@@ -143,12 +143,7 @@
                                     </a>
                                 }
                             }
-                            @if ((Tree.Details?.Lud16Id ?? Tree.Details?.Lud06Url).IsNotNullOrEmpty())
-                            {
-                                <a role="button" @onclick="@Pay" @onclick:stopPropagation="true" class="d-inline-block lh-1 p-2 rounded-5 fs-6 text-secondary me-4 bg-primary-opacity-10-hover">
-                                    <i class="bi bi-lightning-charge"></i>
-                                </a>
-                            }
+                            <PayButton Id="@Tree.Note.PublicKey" OnClick="@Pay" />
                             <a role="button" @onclick:stopPropagation="true" class="d-inline-block lh-1 p-2 rounded-5 fs-6 text-secondary me-4 bg-primary-opacity-10-hover" data-bs-toggle="dropdown" aria-expanded="false">
                                 <i class="bi bi-three-dots-vertical"></i>
                             </a>
@@ -185,7 +180,6 @@
 
     protected override void OnInitialized()
     {
-        accountService.AccountDetailsChanged += AccountDetailsChanged;
         feedService.NoteUpdated += NoteUpdated;
     }
 
@@ -203,13 +197,6 @@
         {
             jsModule = await jsRuntime.InvokeAsync<IJSObjectReference>("import", "../NoteViewer.js");
         }
-    }
-
-    private void AccountDetailsChanged(object? sender, (string accountId, AccountDetails details) data)
-    {
-        if (data.accountId != Tree.Note.PublicKey) return;
-        Tree.Details = data.details;
-        InvokeAsync(() => StateHasChanged());
     }
 
     private void Reply()
@@ -231,7 +218,7 @@
 
     private void Pay()
     {
-        OnPay?.Invoke(Tree.Details);
+        OnPay?.Invoke(Tree.Note.PublicKey);
     }
 
     private void Delete()
@@ -296,7 +283,6 @@
         {
             OnComponentUnload(this);
         }
-        accountService.AccountDetailsChanged -= AccountDetailsChanged;
         feedService.NoteUpdated -= NoteUpdated;
         if (jsModule is not null)
         {
